// core/src/data-table/main/foundation/table.api.ts
export interface TableDragDropOptions<
	RowType = unknown,
	DragItem = TableRenderPropsType.DragObject<RowType>,
	DropResult = TableRenderPropsType.DropResult<RowType>,
	HoveredItem = TableRenderPropsType.HoveredObject<RowType>
> {
	/**
	 * Whether the row can be dragged or not.
	 */
	canDrag?(params: { dragItem: DragItem }): boolean;

	/**
	 * Handler when the row is being dropped.
	 */
	onDrop?(params: { dragItem: DragItem; dropResult: DropResult }): void;

	/**
	 * Whether the target can be dropped or not.
	 */
	canDrop?(params: { dragItem: DragItem; hoveredItem: HoveredItem }): boolean;

	/**
	 * Handler when the dragging begins.
	 */
	onBeginDrag?(params: { dragItem: DragItem }): void;

	/**
	 * Handler when the dragging ends.
	 */
	onEndDrag?(params: { dragItem: DragItem; dropResult: DropResult | null }): void;

	/**
	 * Defines the kinds of dragItems this dropTarget accepts.
	 */
	acceptType?: string;
}